feat: Add 'coming-soon' status for new editions
Introduces a 'status' field to the Edition type, allowing editions to be marked as 'coming-soon'. This change also updates the `Editions` and `OrderForm` components to conditionally disable or hide elements related to editions that are not yet released. Additionally, Issue #6, "Stardust the Super Wizard", is added with the 'coming-soon' status.