feat: Add Chocantes Histórias #01 edition
Introduces the "Chocantes Histórias #01" edition to the Public Domain Fanzine content and updates the UI to display it as a special series. Includes new data for the edition and refactors the Editions component to handle main and special series.

Also enhances the CartModal with WhatsApp integration and payment options, including Pix.